+++ packages/ckeditor5-widget/src/highlightstack.js
@@ -16,7 +16,7 @@ import mix from '@ckeditor/ckeditor5-utils/src/mix';
  * elements. When different highlights are applied to same element correct order should be preserved:
  * * highlight with highest priority should be applied,
  * * if two highlights have same priority - sort by CSS class provided in
- * {@link module:engine/conversion/buildmodelconverter~HighlightDescriptor}.
+ * {@link module:engine/conversion/model-to-view-converters~HighlightDescriptor}.
  * This way, highlight will be applied with the same rules it is applied on texts.
  */
 export default class HighlightStack {
@@ -31,38 +31,22 @@ export default class HighlightStack {
 	 * Adds highlight descriptor to the stack.
 	 *
 	 * @fires change:top
-	 * @param {module:engine/conversion/buildmodelconverter~HighlightDescriptor} descriptor
+	 * @param {module:engine/conversion/model-to-view-converters~HighlightDescriptor} descriptor
 	 */
 	add( descriptor ) {
 		const stack = this._stack;
-		let i = 0;
 
-		// Find correct place to insert descriptor in the stack.
-		while ( stack[ i ] && shouldABeBeforeB( stack[ i ], descriptor ) ) {
-			i++;
-		}
+		// Save top descriptor and insert new one. If top is changed - fire event.
+		const oldTop = stack[ 0 ];
+		this._insertDescriptor( descriptor );
+		const newTop = stack[ 0 ];
 
-		stack.splice( i, 0, descriptor );
-
-		// New element at the stack top.
-		if ( i === 0 ) {
-			const data = {
-				newDescriptor: descriptor
-			};
-
-			// If old descriptor is present it was pushed down the stack.
-			if ( stack[ 1 ] ) {
-				const oldDescriptor = stack[ 1 ];
-
-				// New descriptor on the top is same as previous one - do not fire any event.
-				if ( compareDescriptors( descriptor, oldDescriptor ) ) {
-					return;
-				}
-
-				data.oldDescriptor = oldDescriptor;
-			}
-
-			this.fire( 'change:top', data );
+		// When new object is at the top and stores different information.
+		if ( oldTop !== newTop && !compareDescriptors( oldTop, newTop ) ) {
+			this.fire( 'change:top', {
+				oldDescriptor: oldTop,
+				newDescriptor: newTop
+			} );
 		}
 	}
 
@@ -70,69 +54,88 @@ export default class HighlightStack {
 	 * Removes highlight descriptor from the stack.
 	 *
 	 * @fires change:top
-	 * @param {module:engine/conversion/buildmodelconverter~HighlightDescriptor} descriptor
+	 * @param {module:engine/conversion/model-to-view-converters~HighlightDescriptor} descriptor
 	 */
 	remove( descriptor ) {
 		const stack = this._stack;
-		const length = stack.length;
 
-		if ( length === 0 ) {
+		const oldTop = stack[ 0 ];
+		this._removeDescriptor( descriptor );
+		const newTop = stack[ 0 ];
+
+		// When new object is at the top and stores different information.
+		if ( oldTop !== newTop && !compareDescriptors( oldTop, newTop ) ) {
+			this.fire( 'change:top', {
+				oldDescriptor: oldTop,
+				newDescriptor: newTop
+			} );
+		}
+	}
+
+	/**
+	 * Inserts given descriptor in correct place in the stack. It also takes care about updating information when
+	 * descriptor with same id is already present.
+	 *
+	 * @private
+	 * @param {module:engine/conversion/model-to-view-converters~HighlightDescriptor} descriptor
+	 */
+	_insertDescriptor( descriptor ) {
+		const stack = this._stack;
+		const index = stack.findIndex( item => item.id === descriptor.id );
+
+		// Inserting exact same descriptor - do nothing.
+		if ( compareDescriptors( descriptor, stack[ index ] ) ) {
 			return;
 		}
 
-		let i = 0;
-
-		while ( stack[ i ] && !compareDescriptors( descriptor, stack[ i ] ) ) {
-			i++;
-
-			// Descriptor not found.
-			if ( i >= stack.length ) {
-				return;
-			}
+		// If descriptor with same id but with different information is on the stack - remove it.
+		if ( index > -1 ) {
+			stack.splice( index, 1 );
 		}
 
-		stack.splice( i, 1 );
+		// Find correct place to insert descriptor in the stack.
+		// It have different information (for example priority) so it must be re-inserted in correct place.
+		let i = 0;
 
-		// Element from stack top was removed - fire `change:top` event with new first element. It might be `undefined`
-		// which informs that no descriptor is currently at the top.
-		if ( i === 0 ) {
-			const data = {
-				oldDescriptor: descriptor
-			};
+		while ( stack[ i ] && shouldABeBeforeB( stack[ i ], descriptor ) ) {
+			i++;
+		}
 
-			if ( stack[ 0 ] ) {
-				const newDescriptor = stack[ 0 ];
+		stack.splice( i, 0, descriptor );
+	}
 
-				// New descriptor on the top is same as removed one - do not fire any event.
-				if ( compareDescriptors( descriptor, newDescriptor ) ) {
-					return;
-				}
+	/**
+	 * Removes descriptor with given id from the stack.
+	 *
+	 * @private
+	 * @param {module:engine/conversion/model-to-view-converters~HighlightDescriptor} descriptor
+	 */
+	_removeDescriptor( descriptor ) {
+		const stack = this._stack;
+		const index = stack.findIndex( item => item.id === descriptor.id );
 
-				data.newDescriptor = newDescriptor;
-			}
-
-			this.fire( 'change:top', data );
+		// If descriptor with same id is on the list - remove it.
+		if ( index > -1 ) {
+			stack.splice( index, 1 );
 		}
 	}
 }
 
 mix( HighlightStack, EmitterMixin );
 
-// Compares two highlight descriptors by priority and CSS class names. Returns `true` when both descriptors are
-// considered equal.
+// Compares two descriptors by checking their priority and class list.
 //
-// @param {module:engine/conversion/buildmodelconverter~HighlightDescriptor} descriptorA
-// @param {module:engine/conversion/buildmodelconverter~HighlightDescriptor} descriptorB
-// @returns {Boolean}
-function compareDescriptors( descriptorA, descriptorB ) {
-	return descriptorA.priority == descriptorB.priority &&
-		classesToString( descriptorA.class ) == classesToString( descriptorB.class );
+// @param {module:engine/conversion/model-to-view-converters~HighlightDescriptor} a
+// @param {module:engine/conversion/model-to-view-converters~HighlightDescriptor} b
+// @returns {Boolean} Returns true if both descriptors are defined and have same priority and classes.
+function compareDescriptors( a, b ) {
+	return a && b && a.priority == b.priority && classesToString( a.class ) == classesToString( b.class );
 }
 
 // Checks whenever first descriptor should be placed in the stack before second one.
 //
-// @param {module:engine/conversion/buildmodelconverter~HighlightDescriptor} a
-// @param {module:engine/conversion/buildmodelconverter~HighlightDescriptor} b
+// @param {module:engine/conversion/model-to-view-converters~HighlightDescriptor} a
+// @param {module:engine/conversion/model-to-view-converters~HighlightDescriptor} b
 // @returns {Boolean}
 function shouldABeBeforeB( a, b ) {
 	if ( a.priority > b.priority ) {
@@ -145,7 +148,7 @@ function shouldABeBeforeB( a, b ) {
 	return classesToString( a.class ) > classesToString( b.class );
 }
 
-// Converts CSS classes passed with {@link module:engine/conversion/buildmodelconverter~HighlightDescriptor} to
+// Converts CSS classes passed with {@link module:engine/conversion/model-to-view-converters~HighlightDescriptor} to
 // sorted string.
 //
 // @param {String|Array<String>} descriptor
@@ -159,8 +162,8 @@ function classesToString( classes ) {
  *
  * @event change:top
  * @param {Object} data Additional information about the change.
- * @param {module:engine/conversion/buildmodelconverter~HighlightDescriptor} [data.newDescriptor] New highlight
+ * @param {module:engine/conversion/model-to-view-converters~HighlightDescriptor} [data.newDescriptor] New highlight
  * descriptor. It will be `undefined` when last descriptor is removed from the stack.
- * @param {module:engine/conversion/buildmodelconverter~HighlightDescriptor} [data.oldDescriptor] Old highlight
+ * @param {module:engine/conversion/model-to-view-converters~HighlightDescriptor} [data.oldDescriptor] Old highlight
  * descriptor. It will be `undefined` when first descriptor is added to the stack.
  */